Move all project from ContentView to ViewListOfNames
